Simplify MetaPlaylist Manifest loading by relying on the same code than other transports
I was PoCing protocol-detection when initially parsing the Manifest, so
application developers wouldn't e.g. have to tell to us that they want
to play "dash".

While doing that, I noticed that all transports had close to the same
logic for Manifest fetching. Only differences I've seen are:

  - DASH may add integrity checks if the `checkManifestIntegrity` option
    is set

  - DASH may ask the browser to obtain the response as an ArrayBuffer
    (and not as a default JS string) if it is probable that we will rely
    on the WASM MPD parser

  - the `local` transport throws if no custom `manifestLoader` is
    declared.

As such, those differences look minor and straightforward enough so the
corresponding logic can be factorized into e.g. a single parameterized
`createManifestLoader` function.

This was already done for DASH and Smooth, but I don't know why we
didn't use it for the experimental MetaPlaylist feature. This commit
fixes this.

We could also use it for the experimental `local` transport in the
future, though I'm unsure of how we could generically make sense of its
peculiar rule (no real loading).