[POC} Only load media segments for Representations known to be decipherable
Historically, the RxPlayer load media segments and perform the
license-fetching logic in parallel, as it theoretically lead to faster
playback when all qualities are decipherable.

However, this has always only been a theory and it may have the reverse
effect on devices whose some keys are not obtainable (this also is only
theoretical).

To better know what the impact of loading media segments in parallel or
after the license, I'm here only loading media segments (not the
initialization segment, which can be loaded right away) either if the
Representation is clear or is known to be decipherable.

This is only a Proof-Of-Concept with for now no intent to make it a
feature.