Remove reliance on the rimraf dependency
We often have to work on debugging applications. In most of those cases,
we call `npm-link` or `yarn-link` to link our local version of the
RxPlayer to a given application.

Sadly it turns out that some of those applications have complex and
poorly-configured package installation and build steps, which leads to
many errors as we try to produce a build.

An issue I was having right now, is linked to a peculiar way packages
can be installed in one application at Canal+, leading in some cases to
issues with how the `rimraf` package is imported.

I grew tired of this issue. Considering that the `rimraf` dependency
for our simple usage seems to be very easy to replace with a few line of
Node.js (fs.rm fills what we want to do and seems to even be relied on
by `rimraf` on the simple cases we're exploiting), I decided to do just
that - so there's less possibility for dependency-importing issues to
occur.