Work-around firefox 129 windows bug
We noticed a pretty big issue on firefox 129 (released yesterday) on
windows) where we cannot play encrypted content anymore (seen on
production with multiple applications relying on the RxPlayer - so it's
a relatively big issue).

It seems that since this release, the
`navigator.requestMediaKeySystemAccess` EME API, which we use to poll
available decryption capabilities in the current environment, will fail
as soon as there's at least one asked "robustness" (allows for example
to ask for hardware-based DRM) even if we're asking for e.g.
software-based DRM, which we know the device has.

It seems likely this is a firefox issue, but it seems too big for us to
just do nothing and throw the bug at them.

Other streaming actors, such as Netflix and Amazon Prime Video don't
seem to be impacted by the issue at first glance.
In fact, they are, but they have a fallback call where they are calling
the `navigator.requestMediaKeySystemAccess` API without any robustness
just in case the API is broken like here.

Yet, we also have the same work-around, but we integrated both
configurations (the first with robustnesses, the second without), in the
same `navigator.requestMediaKeySystemAccess` call, whereas they're doing
two calls each with an array of a single element (the first with, the
second without like us).

So to """""fix""""" the issue, I decided to align with their
work-arounds (so the next time it breaks, browser devs will see it also
breaks Netflix and thus maybe detect the issue faster).

For the same key system type, doing one call with multiple configurations
or multiple calls with one configuration seems equivalent from my
understanding of the EME recommendation, so what I did here should
change nothing EME-wise (except more EME API calls), it is just to align
the code with others.

I still had to change our recently-added code about codec checking, to
make it work robustly in that new scenario.