[v3] Fix auto-play on tizen when it starts on a discontinuity
In one of our recent updates on the initial seek and play, we appeared
to have broken autoPlay on Samsung TVs when they started on a
discontinuity.

This is yet again an issue linked to those infamous Tizen seek-backs I
love to complain about.

The fix proposed here is to just call the HTML5 `play` API even if we
appear to be rebuffering from the amount of buffered data in
front of us on Tizen if it tells us that it actually have enough data
to play through its `readyState` property.

This shouldn't break things and seems logical. I hesitated doing the
same fix for all targets but the `readyState` property is known to be
sometimes wrong on some devices so I didn't.