Treat codecs beginning with `hev1` as hevc
While working on new DRM integration test in #1478, to add tests for the
not-yet-released features proposed by #1484, I noticed that some
contents anounced HEVC by beginning the codec string with `hev` and
others with `hvc`.

When comparing family of codecs, we thus had false negative which may
mean we would have ended up playing encrypted hevc on devices where it
wasn't supported, if it was anounced in the MPD as `hev1...`.

I now treat both the same way.