[Proposal] Add `filterPlayableRepresentations` property to audio and video tracks API
Initial reason for this proposal
--------------------------------

I was prototyping an heuristic for an application where it would
estimate the maximum reachable quality due to DRMs on the current
device.

The idea is that the RxPlayer is (at least for now) not able to guess
the security policy of provided contents in advance (before actually
communicating the license to the CDM) whereas the application may have
a better idea (e.g. most 1080p+ contents from some identified provider
may have the same drastic security conditions, so if a quality is not
playable for one content, it probably won't work for another).

Until now, the solution proposed by the RxPlayer - which also works - was
to first test all qualities for all contents and then fallback from
undecipherable qualities once they become known.

However, doing this at each content load may, on the more restricted devices
(mostly PCs and mobile phones when on the web), mean that we'll at each load
attempt to play higher qualities just to fallback to a lesser quality when
we see that we cannot handle higher ones.
This fallback isn t usually perceptible much by the customer, yet up to a few
hundred of milliseconds may be "lost" on each load that way (median of ~150ms
on my PC on widevine L3 on frequently-encountered policy levels and provided
qualities).

So to improve on this situation the idea is to be able to determine
application-side when qualities become undecipherable (knowable through
`representationListUpdate` events) and which have become undecipherable
(no method until now could do that, you had to awkwardly diff between the
list of qualities at load time and after a `representationListUpdate`
event).

Then the application is able to limit the maximum played quality in further
content loading through the `lockVideoRepresentations` method.

Because decipherability may evolve on a same device over time (e.g.
linking / unlinking an external monitor may change the possibility to
rely on the HDCP security mechanism), this algorithm on the
application-side still may have to re-check periodically which qualities
can or cannot be played. Though in the seen usage, this is simple to do
(for example, this is mainly automatic on a `singleLicensePer:
"content"` or `singleLicensePer: "periods"` mode - as we'll know the
decipherability of higher qualities even without playing them, for the
default mode, a more advanced strategy of trying the higher quality once
enough buffer is build is possible).

Proposal
--------

Before, we only listed "playable" representations in our audio video
tracks methods (`getVideoTrack`, `getAvailableVideoTracks` ...) and
events (`videoTrackChange`, `availableVideoTracksChange` ...).

Now the idea is for the former (the `get` methods), to optionally also
include non-supported Representations (either because in an unsupported
codec, or because undecipherable, with this reason explicited through
properties returned by those APIs) if a supplementary
`filterPlayableRepresentations` boolean has been set to `false`.

Though the reason I implement this may be seen as somewhat niche, the
possibility to obtain metadata all Representations, including the ones
that are not actually playable, may also be useful if the application
wants to be able to debug decipherability issues programatically (e.g.
polling from its customers how many are unable to decipher higher
qualities). So this new API appears to be compatible with multiple
usages.