Fix bug where active Period might never be known
While testing for an upcoming release, we found a bug where an
application wasn't able to tell which period was currently-played and
thus wasn't able to show the audio and text track list for the current
Period to its user.

Turns out that the issue at its base, is due to my poor comprehension of
the `Array.prototype.forEach` JS API. I though that it operated on a
sliced/cloned array but it actually appears to re-consider the source
array at each iteration, raising internally an index while doing so,
until the end of the array is reached.

What I mean is that:
```js
var a = [0, 1]
a.forEach((b) => {
  console.log(b);
  a.splice(0, 1);
});
```

Logs:

```
0
```

Where I thought it would have logged:

```
0
1
```

(I thought mutations on the source array where not influencing the
forEach call because I thought for whatever reason that it initially
cloned it. But no, it does not).

I hope this is the only place where that trick was done. I'll have to
review all those iterator methods just to check!

There's also weird things happening with other operators such as
`Array.prototype.map` where the initial length of the source array
becomes the length of the destination array, even if the source
array's length is updated inside map's callback, we may thus end-up
with "empty" slot in the resulting array if the source array is
shrinked (and missing on the new values if the source array grows).