Simplify Representation's `keyIds` metadata
`Representation` objects allow in the RxPlayer code to store the
metadata linked to a given [audio, video or text] media quality.

One of its properties `contentProtections`, allowed to store DRM-related
information linked to that quality.

Amongs its two properties, `keyIds` was in my opinion unnecessarily
complex: it contained not only the key-id(s), as `Uint8Array`(s), but also
optionally the corresponding `systemId` (an identifier for the DRM scheme
to which that key id is related) for each.

This doesn't appear to be necessary nor make much sense in most contexts:
a key id is supposed to uniquely identify a given decryption key, and
the key id is generally the same regardless of the system/technology, at
least if the same decryption key is relied on (which probably should, as
for now in all practical cases the same source data will then be decrypted
by all systems).

Even in purely theoretical cases where multiple key ids would exist
depending on the key system for whatever reason or where different parts of
the media data would be decrypted depending on the key system, I suppose
that no conflict should exist where a key id in one key system would lead
to a different key in another key system - meaning that that metadata
(which is mainly used as an identification mechanism) should not conflict.

If for whatever reasons we do encounter such case in the future, we may
re-complexify the logic.