[Proposal] Add `contentProtections` to the tracks API
In the same spirit than #1501, I found out that an application had no
practical way (beside the `representationFilter` API) of knowing which
`Representation` (quality) was encrypted when calling video and audio
track APIs.

Unlike #1501, I thought that this information could always be delivered
without the need of a supplementary API.

Now comes the question of how to format that metadata: An `isEncrypted`
key? A `contentProtections` object?

For now, I chose to align with the `representationFilter` API, where we
have a `contentProtections` object at the Representation-level, which is
either set to `undefined` if unencrypted or if we don't know if it is,
or to an object with a `keyIds` property if it is.

That `keyIds` property is set as an array of `Uint8Array`, corresponding
to the key id(s) linked to that Representation.

For now I chose not to also include DRM `initialization data` (e.g. the
ISOBMFF's pssh box) in that object to let us the possibility to
internally work on its format without having to re-format it in a
specific way for the API.