DRM: Re-create MediaKeys for each content on Philips' NETTV
Fixes #1464

There's a new device (after LG and Panasonic TVs) which seems to have
issues when relying on the same `MediaKeys` instance on multiple
contents, which seems to be Philips's NETTV (both 2023 and 2024 devices
seem to have the issue as far as we know, and they advertise completely
different operating systems, so instead we identified the culprit as
"NETTV", which seems to be the web browser application Philips developed
for their TVs and which is present on web applications' user agent on
those devices).

So I disable the possibility for them here.
I still have to check if it works because I took as user-agent something
I found a little randomly on the web.

Note that disabling re-usage of the `MediaKeys` instance will mean no
in-memory `MediaKeySession` caching, which is a trick we use to both
speed up loading a previous content and preventing to much requests
on the license server also when doing that.

But we know most other players don't seem to do such tricks and the
speed up is not that much noticeable, so that shouldn't be a big issue.