[Proposal] export a build script to facilitate rx-player tests
Quick description
-----------------

This commit is a proposal to allow applications using the RxPlayer to
easily depend on an RxPlayer development branch or any commit pushed to
Github directly in their package.json, which is as easy as adding:
```json
    "rx-player": "canalplus/rx-player#<GIT_BRANCH_OR_SHA_1>",
```
To the `package.json` (instead of the usual version).

_Almost too easy, surely this has nothing to do with npm being a Github
subsidiary._

This in turn will allow them to very easily test future features, fixes
and improvements without our involvement.

Previous situation
------------------

Doing this previously was never really easy in the end because we don't
push builds to GitHub, and so npm/yarn, in the applications' project,
would just pull our source code and then don't know what to do with it,
ultimately throwing on unfound files when they try to build their
project.

We, RxPlayer developers, relied on that trick as we knew how to
navigate that type of situation but didn't ask application developers to
do so because the work-arounds to do depended on the project and were
sometimes arcane (updating things in the CI, circumventing a local npm
install etc.).

Instead, when we had to ask developers to test, we had several solutions
on a case-by-case basis, yet none of them being not perfect.

The added `clean_build.mjs` script
----------------------------------

This commit adds a file `scripts/clean_build.mjs` which doesn't require
any dependency besides node ones.
It dynamically calls `npm install` and then once done build the
rx-player in place, by importing the right script.

There are some tricks because node/npm/yarn like doing black magic, one
of the main one is an issue we had for a long time on one of Canal+'s
application: For some reasons, they depend on a very old version of a
dependency (from 2017) which for some other very weird reason has an
hard dependency on `npm` (also one from 2017, a literal dinosaur).

Because of this, a very old version of `npm` is installed in their
`node_modules` directory, and thus when running the `clean_build.mjs`
script from a `package.json`'s script (which is the easiest way to run
it), the `npm` command actually call that very old version which
somehow do not succeed to install our dependencies (one of ours must
have a nasty poorly compatible `postinstall` script I guess :p).

So I added here a kind-of ugly work-around: if `npm --version` seems
to return a very old version, we check if we're not relying on a local
npm. If we are (there's `node_modules` in the path), we use `whereis`
to use another one if it exists. This is complex magic thing I don't
like to do, but when the opponent also rely on magic I have few
choices left.

How to run this
---------------

At first, I though I could just add a `postinstall` script inside the
RxPlayer, which would check if the builds are made on the pulled
dependency, and if not, produce them.

It works and would have made depending on a git branch **VERY** easy,
but I thought it wasn't really polite to execute on all applications
pulling us some random scripts - which could have bugs - and which ran
some executable (`npm`, `whereis` etc.) on the host.

So I changed my mind and chose to expose the `clean_build.mjs` script
from the `"rx-player/scripts/build"` path instead.

The idea is that an application will have to iself create some script
(doable from a `package.json` `postinstall` script), which will import
that file and call the build function from it.

This is a little more involved, even more due to the CommonJS/ES module
shenanigans which kind of forces us to create a temporary file in the
meantime with an `.mjs` extension, but everyone should be able to do it.

Thus I chose to add instructions in our `README.md` to tell people how
to run this, the idea is that we would redirect people both to the
git hash to pull and to that `README.md` chapter to tell them how they
can test it in their application.