CMCD: Implement `nor` and `nrr` CMCD keys
Based on #1467

We're planning to add CMCD (Common Media Client Data) v1 capabilities to
our next release of the RxPlayer v4.

This features adds multiple key-value pairs with most request (either
from the query string, either from headers) to inform the server-side of
the current playback condition from the player which made the request.

We were previously missing two keys used for a specific use-case: `nor`
(Next Object Request) and `nrr` (Next Range Request). Those two allow
the player to anounce which segment it will probably load next so that
the server could theoretically prefetch it in advance.

Implementing those two means that our CMCD logic needs to be aware of
the next segment that will probably be requested after a current one.
This is made much simpler by #1467, as now our fetcher logic is aware of
the whole queue of segments for a given type (audio/video etc.) in most
cases.

The main meat of this PR is just computing a relative URL, because
that's what CMCD wants (in the name of DoS mitigation, I did not dive
into why). I thought we already had that logic somewhere but if we have,
I did not find it, so I re-coded and tested it.