Fix issue when reloading while a reload is pending.
I noticed of an issue with the `MULTI_THREAD` feature, where an
application would receive a `CancellationError` in the very rare
(possible right now?) occurrence where we're reloading while a reload
operation is already pending.

We don't ever want to leak `CancellationError` which is just an internal
mechanism. I thought that we can just ignore that one.