Add A1 KSTB 40XX to our list of devices which cannot reuse MediaKeys
A partner encountered new cases where multiple content loadings lead to
issues, that disappeared if we re-create the `MediaKeys` instance at
each load.

There may be a common denominator (like common software) between the
raising list of devices on which we see the issue recently, but we did
not pinpoint it for now, here so I only detect the specific device and
perform the work-around on it and the other devices on which it has been
seen for now.