[Proposal]: Add experimental.enableResolutionChecks loadVideo option
This is a Proposal to optionally move some "resolution checks" that
were historically on the application-side on the RxPlayer side.

Some devices aren't able to play video content with a higher resolution
than e.g. what their screen size can display: some will fail on error,
others will just infinitely freeze etc.

Many of those devices have vendor-specific API that indicated whether a
device was e.g. compatible to 4k content, 1080p content etc.

Until now we told the application developers to do the Representation
filtering themselves with the help of the `representationFilter` option
from the `loadVideo` API. This is because of the vast numbers of devices
out there, each with their own API, and because we didn't want to take
the risk of having false negatives if it turned out some of those
API were not always right.

However, many of those devices are very popular (Lg and Samsung TVs,
game consoles). Thus I'm wondering if it would be better that we
provide some of those resolution checks ourselves, to lower the
efforts an application have to make to rely on the RxPlayer on those
common devices.

For now I added an experimental `loadVideo` option:
`experimental.enableResolutionChecks`, that has to be set to `true`
to enable the behavior. The long term idea would be that for devices
where it seems 100% accurate, we would enable the check by default.

It will directly filter out resolutions that are too high, unless all
resolutions are too high on the current video track in which case it
will disable the check (as a security).
In other cases, it follows the exact same rules than the
`isCodecSupported` and `decipherable` properties of `Representation`.