Fix adding RxPlayer to the global scope for bundles
I recently noticed while doing quick tests that the bundles we attach
to each release note (and publish to npm, even though that's not what is
used when importing the RxPlayer) was not exposing the `RxPlayer` to
`window` - which I suppose is how most people would be interacting with
that bundle.

It seems that the v4.1.0 release note for example has such a bundle right
now (maybe I'll update it there, though bundles published to npm cannot be
retroactively updated).

The issue might have appeared when we removed the webpack dependency,
there was an option in our bundling script to expose the RxPlayer but it
was never set - thus the fix here is pretty simple.

PS: On that matter, I'm not sure publishing bundles to npm serve a role,
as people relying on the `npm` registry are most likely relying on
`import`/`require` statements and do not just manually copy one of the
bundle files that would be downloaded after calling `npm install` / `yarn`.
Bundles would still be accessible through release notes.
Though it seems many people still rely on bundles (for simplicity
reasons?), so maybe people actually DO that? (no idea how they get that
bundle here)