[v3] [Proposal] Bring prettier to the v3
One of the bigger pain points of backporting fixes and improvements to
the v3 is due to the fact that we rely on prettier for the v4 but not on
the v3, leading to huge conflicts in the end mostly related to different
formatting.

This PR proposes that we bring prettier and a similar eslint config to
the v3 just to ease comparisons between both major versions.

Note that some eslint rules enabled on the v4 are here not enabled,
mostly `eqeqeq` and `no-nested-ternaries`(both because that's a lot of
work).

For that second rule, the fact that we also run prettier may lead to code
harder to read than before.