Add buffer size estimate to debug buffer content graph
While continuing proof-of-concepts related to how much buffer we build,
I found that our estimate of the current buffer size in bytes (more
exactly the combined size of pushed segments that still have at least a
sub-part present in the buffer) was a potentially-interesting metric
that we did not expose anywhere: neither in logs nor in our
DEBUG_ELEMENT feature.

This commit adds it to the `DEBUG_ELEMENT` feature, just after the
corresponding buffer content graph.

It may be even more useful in logs, but I did not bother to do it for
now as I currently mainly relied on the debug element.