Fix MSE queue blocked in some situations
I noticed that an error thrown (and not a promise rejected) by the
MSE `appendBuffer` and `remove` API did not lead to an `updatestart` nor
an `updateend` events which were our triggers to start the next MSE
operation in our logic.

If we were unlucky to have future operations waiting (in reality
this situation is kind of rare as consecutive push operations are
merged together) we would be left with a queue only being filled
and never being emptied.

To reproduce it, I had to play a local content with a huge buffer size
and seeking between DASH Periods like crazy, so hopefully this seems
not easily reproducible, but it may well be on non-powerful devices
with a too high buffer size?

As written in the code, the solution comes with caveats, but it's still
better than what we had before.